pointerevents/webidl_pointerevent.html

<!DOCTYPE HTML>
<html>
<!-- 
     Test cases for Pointer Events v1 spec

     These test test the Web IDL of the PointerEvent event.

     This document references Test Assertions (abbrev TA below) written by Cathy Chan
     http://www.w3.org/wiki/PointerEvents/TestAssertions
-->

<head>
  <title>PointerEvent Web IDL test</title>
  <meta name="viewport" content="width=device-width">
  <script src="http://w3c-test.org/resources/testharness.js"></script>
  <script>
    setup({explicit_done: true});

    // Check a PointerEvent's atrributes for conformance to WebIDL: attribute exists,
    // has proper type and if the attribute is "readonly", it cannot be changed.
    // TA: 1.{1,2}
    function check_PointerEvent(ev) {
        test(function() {
           assert_true(ev instanceof PointerEvent, "event is a PointerEvent event");
        }, ev.type + " event is a PointerEvent event");

        [
          ["readonly", "long", "pointerId"],
          ["readonly", "long", "width"],
          ["readonly", "long", "height"],
          ["readonly", "float", "pressure"],
          ["readonly", "long", "tiltX"],
          ["readonly", "long", "tiltY"],
          ["readonly", "DOMString", "pointerType"],
          ["readonly", "boolean", "isPrimary"],
        ].forEach(function(attr) {
            var readonly = attr[0];
            var type = attr[1];
            var name = attr[2];

            // readonly check
            if (readonly === "readonly") {
                test(function() {
                    assert_readonly(ev.type, name, ev.type + "." + name + " cannot be changed");
                }, ev.type + "." + name + " is readonly");
            }

            // existence check
            test(function() {
                assert_true(name in ev, name + " attribute in " + ev.type + " event");
            }, ev.type + "." + name + " attribute exists");

            // type check
            switch(type) {
            case "boolean":
                test(function() {
                    assert_equals(typeof ev[name], "boolean", name + " attribute of type boolean");
                }, ev.type + "." + name + " attribute is of type boolean");
                break;
            case "float":
                test(function() {
                    assert_equals(typeof ev[name], "number", name + " attribute of type float");
                }, ev.type + "." + name + " attribute is of type number (float)");
                break;
            case "long":
                test(function() {
                    assert_equals(typeof ev[name], "number", name + " attribute of type long");
                }, ev.type + "." + name + " attribute is of type number (long)");
                break;
            case "DOMString":
                test(function() {
                    assert_equals(typeof ev[name], "DOMString", name + " attribute of type DOMString");
                }, ev.type + "." + name + " attribute is of type DOMString");
                break;
            default:
                break;
            }
       }); 
    }
    
    function run() {
      var target0 = document.getElementById("target0");

      // Set up a timeout so that if the UA doesn't support PointerEvents,
      // the framework will report "NOT RUN"
      var test_pe = async_test("This User Agent does Not support PointerEvents.");

      on_event(target0, "PointerDown", function onPointerDown(ev) {
          ev.preventDefault();
          check_PointerEvent(ev);
      }, "PointerDown tests OK ...");
    }
</script>

<style>
  div {
    margin: 0em;
    padding: 2em;
  }
  #target0 {
    background: yellow;
    border: 1px solid orange;
  }
</style>

</head>
<body onload="run()">
  <h1>PointerEvents: check Web IDL of PointerEvent event</h1>
  <div id="target0">
    Touch this box with one finger (or other pointing device) ...
  </div>

  <div id="log"></div>
</body>
</html>
